+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 
 ## Unreleased changes
 
+### Changed
+
+* Missing repos no longer appear as search results.
+
 ### Fixed
 
 * In Sourcegraph Data Center, searches no longer block if the index is unavailable (e.g., after the index pod restarts). Instead, it respects the normal search timeout and reports the situation to the user if the index is not yet available.